Moved FormResult classes to their own files. The Winforms designer tool in VS does not work when Winforms classes have other classes declared in the same file.
